Create OnFolderMove
When a content item is moved (Folder Actions - Move) to a new folder , the security group of the content item is not updated to the one set for target folder metadata. 
This idoc script will set the content item's security group to the one that is set for Folder (target). 

Add this idoc script in derived section for dSecurityGroup metadata .