Append propagated ACL values with existing ACL.
Framework Folders has propagate functionality where in a set of metadata updated can be set to all the content items / sub-folders under it. The drawback with this is when propagate is run it overwrites all the existing ACL fields (User access list , role access list or group access list) of the content items and only the new values are saved . Previously existing values are cleared .
This idoc script rule will append the existing ACL values with the ones being propagated there by maintaining the older and new set of access values.
The script is set under "Derived Value" for xClbraUserList (this example takes in consideration for user access list).